Create README.md
What if I told you that we could hack time to recover over $3 million in Bitcoin from a software wallet that's been locked since 2013? In this episode, I join forces with my friend Bruno to reverse engineer the RoboForm password generator in order to regenerate passwords that have been generated in the past.